requirements.txt
python-telegram-bot==20.7
dateparser>=1.2
pytz>=2024.1
tzdata>=2024.1
gspread>=6.0.0
google-auth>=2.30.0

